requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.6.20
cfgv==3.1.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==2.2.14
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.0.0
django-sendsms==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-user==0.0.8
drf-yasg==1.17.1
drfaddons==0.1.3
filelock==3.0.12
identify==1.4.21
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.4.0
packaging==20.4
pre-commit==2.5.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.15.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.25
wrapt==1.12.1
zipp==3.1.0
